requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap-modal-forms==2.0.0
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dotenv==0.10.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
django-solo==1.1.3
wheel==0.35.1
django-colorfield==0.3.2